Add check_fix for MisconfiguredRedirection
Summary:
# Context
Eden doctor is a tool used to try and fix eden when problems arise.

# This Diff
This diff implements the check_fix for MisconfiguredRedirection and tests

# Technical Details
Redirections are detected in the following method
Bind mounts
- All mounts are enumerated. All mounts starting with the checkout root excluding the root itself are added to a temporary list.
- All configured redirections are enumerated. All bind mounts that are also configured are correctly configured. Otherwise they are Unknown Mounts
Symlinks:
- All configured symlinks that resolve correctly are Configured Correctly. Otherwise they are Incorrect. If there is an OsError, then they are Missing.

The fix is to remove the mount/symlink and recreate it if it is not an Unkown Mount.

The check for the fix is to detect if an Unknown Mount is removed, and otherwise if the bind/symlink exists.

Reviewed By: genevievehelsel

Differential Revision: D59821474

fbshipit-source-id: 16e0880c6c2c2c16f087a22c553ff67efb149667