move the core parts of interest handle ops
Summary:
In [this diff] (https://www.internalfb.com/diff/D59356329) we introduced a new cache ScmStatusCache which should be protected by a lock when updating the entries. We created it based on ObjectCache which has the ability to evict when it reaches the size limit.

However, `ObjectCache` itself has an internal cache which protestcs the core operations. This lead us into a world of double locks.

We are trying to resolve this by refactoring out `ObjectCache` into two classes: thread-safe one and thread-unsafe one. Then we can make the original `ObjectCache` a subclass of the thread-safe one but `ScmStatusCache` a subclass of the thread-unsafe one.

This diff serves as a first attempt to refactor some functions of `ObjectCache` so we can easily play with the logics which need to be protected under a lock.¡¡

Reviewed By: kmancini

Differential Revision: D59737795

fbshipit-source-id: b34d89a986283098fa3ed98d5b6393b7f72d9a9e