Extend integration test with valid backsyncing case
Summary:
## This stack
Implementing support to backsync changes that touch the submodule expansions.

# This diff
Extends the big integration test with an important case: backsyncing a valid change to the submodule expansion.

This is currently not supported and we block any change that touches a submodule expansion from being backsynced.
But D60118876 will introduce support for backsyncing these changes, so this diff already adds this scenario.

Differential Revision: D60372665

fbshipit-source-id: a12e83d7b8ede1fb8d1e17c400540079ccf43016