Introduce type to ensure submodule validation has been performed
Summary:
We'll soon introduce support for backsyncing **valid** changes to the submodule expansion.

Most of the implementation will **rely that the provided bonsai has valid submodule expansions**, i.e. that validation ran and was successful for it.

So let's introduce some typesafety to ensure this.

This introduces a wrapper type without a public constructor. The only way to instantiate it is by running submodule expansion validation.

This type will be the input of all functions that require valid bonsais, so **it's safe for them to make assumptions**.

Differential Revision: D60457309

fbshipit-source-id: 28c90ae65d1db20dc8013a0de3061e0a5feea423