walker: add logic for querying new table for checkpoints
Summary: The walker relies on a `walker_checkpoints` xdb table to keep track of what changesets it has already walked. This table stores ranges of ids that it gets from the `changesets` table. The `changesets` table is deprecated and replaced by the `commit_graph_edges` table, so we will need the ids to reference that instead. To enable this migration I'm creating a new `walker_checkpoints_v2` table that has the same exact schema and semantics, and gating it behind a command line argument.

Differential Revision: D60454065

fbshipit-source-id: 4e40ba918cdd2c21e18f43a9fec520d3e864738e