Honor Source of Truth when accepting push in Mononoke Git server
Summary: We now have a source of truth flag per repo that is stored in a DB table. This table will be used by both Metagit and Mononoke to decide where git pushes should be accepted. This diff makes the necessary changes so that Mononoke doesn't accidentally accept pushes when its not source of truth

Differential Revision: D60473115

fbshipit-source-id: 75f7ae00f83cd2e062b46849168967b649b2bab3