Rename HookFileContentProvider to HookStateProvider
Summary: The trait `HookFileContentProvider` was created as a way to access necessary information that is outside of the hook interface, e.g. file changes. However, now the name is misleading since it doesn't just provide content but also provides metadata, directory information and bookmark state. This diff renames it to something slightly more accurate.

Reviewed By: mitrandir77

Differential Revision: D60513719

fbshipit-source-id: 59a773acc49eeae4d52a394bb9cf08248bc0a3c4