Provide mechanism to access bookmark state in ChangesetHooks
Summary: Initially, I planned to provide this functionality by adding a separate set of facets within `HookManager` but then realized that `RepoHookFileContentProvider` already has access to the `Bookmarks` facet so extended the `HookFileContentProvider` trait to allow access to the bookmark state. This should unblock the hooks that could not get migrated from Git to Mononoke so far.

Reviewed By: mitrandir77

Differential Revision: D60513361

fbshipit-source-id: 99b3a678781b1a6fb50182bb58dbcb840f607707