Integration test highlighting git push failure in case of ref deltas
Summary: When the server has some large object (tree or blob) that's been modified in the client's repo, instead of sending the entire modified object as part of the push, the Git client sends a ref delta to the large object thereby saving on bandwidth. The current handling of ref delta in Mononoke Git server during push path is incorrect due to which the push fails. This integration test highlights the failure. The next diff will fix this problem.

Reviewed By: singhsrb

Differential Revision: D60531206

fbshipit-source-id: a67735f6f07cb5c74d33482994a6a89a6fdb7a1e