e-menu usage telemetry (1/4): Add a new `EMenuActionEvent` in LogEvent.h
Summary:
## Background:
E-Menu is a system tray icon that will help you to know more about your checked out Eden repos, to run some operation using Eden Daemon and to report any issue with the Eden. There is no way today to track usage of the usage of this icon.

## Solution:
Introduce a new Scuba event for e-menu usage and intrument it for various user interactions with e-menu starting with e-menu click

## Implementation:
1. **Add a new `EMenuActionEvent` in LogEvent.h**
2. Introduce `StructuredLogger` in WindowsNotifier
3. Add `logEvent` method to WindowsNotifier to intrument events to Scuba
4. Call `WindowsNotifier.logEvent() with `EMenuClick` event when system tray icon is clicked

## Tracking:
[Scuba](https://fburl.com/scuba/edenfs_events/3bbv9j15)

Reviewed By: MichaelCuevas

Differential Revision: D60459465

fbshipit-source-id: 7e52e26281f4c7f71d7a564b6b4f5154eb511fc6