slow_bookmark_mover: replace waiting for segmented changelog with waiting for phases
Summary: The serverside segmented changelog is no longer used, so there's no reason for the `slow_bookmark_mover` tool to have a way to wait for the segmented changelog tailer to catch up. On the other hand, it can be useful to have a way to wait for the phases of commits to be updated as a large bookmark move can otherwise cause a delay in warm bookmarks cache while the phases warmer catches up.

Reviewed By: gustavoavena

Differential Revision: D60518160

fbshipit-source-id: 621edddb8ca1a37184e8124e56534cec88022563