commit_graph: implement CommitGraphBulkFetcher
Summary:
Context: The walker relies on `bulkops::ChangesetBulkFetcher` to iterate over all changesets in a repo. In order to get rid of the changesets facet, we need an equivalent commit graph based way to this.

Implemented a new `CommitGraphBulkFetcher` facet that has the same functionality as `ChangesetBulkFetcher` but uses the commit graph instead. This facet allows returning a stream of all changesets whose auto-increment sql ids lie within a given range.

Reviewed By: gustavoavena

Differential Revision: D60526860

fbshipit-source-id: 3b879f7b3f7f36b0023d1bccda87eea6d4608ec5