Fix insidious invalid packfile error due to optimized streaming
Summary:
In S434542, we realized that the bundles generated by `repo_stack_git_bundle_store` endpoint are sometimes invalid. We were able to correlate this problem with the flip of the JK `scm/mononoke:use_optimized_git_packfile_stream` indicating that the source of the problem was the optimized stream path used for bundle generation. However, we were unable to truly understand the root cause due to:
- The error never appearing again even with the JK on when tried by members of the team
- None of the integration tests show any trace of this error even when run in stress mode

As a last ditch effort, we tried to enable this JK by default for integration tests (see Youssef's previous diff) so that it can be caught by the CI. This worked and the error seemed to surface again but only for `dev-rust-oss` build mode. This led us to believe that the underlying issue could be a timing problem that surfaces when running unoptimized code, which was indeed the case. This diff fixes that problem. I will write a separate post that explains this problem in detail.

Reviewed By: YousefSalama

Differential Revision: D60605642

fbshipit-source-id: c1937f8972d36cda35bdfc2aed1ec699457a0540