Use listfile0:- file pattern to pass long lists of files via stdin to fix too long arguments
Summary:
If you have 1000s of files in your uncommitted changes, then uncheck one of them and click "discard", it would pass all the files to `sl purge`. If you have LOTs, this could be too many for the OS to handle. On MacOS, the limit is like 1,000,000 chars, but it seems it's much lower on windows.

Instead of sending files in as command line arguments, it's better if we can pass them via stdin. This is supported by sl using file patterns (`sl help patterns`) with `listfile0:-`. This gets around the length limit nicely, so we don't have to batch our operations.

We do have a mechanism for Operations to define what their stdin should be. However, we want to not just pass the files, but also normalize them to be repo-relative on the server. Thus I opted for a new CommandArg type which expands into an argument + stdin. This allows the server to process the list of files as needed. It's also easy to write in the definition of the Operation. This required just a little shuffling of how we process args and stdin.

Reviewed By: muirdm

Differential Revision: D60557866

fbshipit-source-id: a751a0b7c2aa74dc5c0d2e6154409dccdcaccd48