walker: modify checkpoint chunk bounds calculation logic
Summary: After the walker processes a chunk of changesets it tries to update a checkpoint to be able to resume processing from after the processed chunk. When fetching changesets, the walker uses `ChangesetBulkFetcher` which internally fetches changesets using batch queries, and then flattens batches into a stream. The confusing thing for me is that the walker uses the bounds of the internal fetched sql batch to determine how to update the bounds of the checkpoint, even though the internal batch could be different from what was processed from the stream because the batches are flattened internally. This diff makes the checkpoint bounds calculation depend on the ids of the actual processed changesets which seems like the logical thing to do me.

Reviewed By: andreacampi

Differential Revision: D60596064

fbshipit-source-id: 38536f3c9cc8d8618adc34a88b311fffa15492c0