git/protocol: don't poll delta_manifest_entries_futures if it's empty in the optimized stream
Summary: Check if `delta_manifest_entries_futures` is empty and don't poll it if so. This doesn't seem to technically be necessary unlike the fix from D60605642, but streams are not supposed to be polled if they ever return `Poll::Ready(None)` and invariants inside of `FuturesOrdered`/`FuturesUnordered` might rely on that so let's be safe.

Reviewed By: RajivTS

Differential Revision: D60651538

fbshipit-source-id: 21b1cf8032b7d6a5427d67ea8eff1b924995dec0