Increase the default SQL cache size
Summary:
The defaults date back from when we had almost no use in production. These days we get some queries about snapshots in SCS, so we are already seeing more misses than hits. When I enable pushredirects, that will probably make the cache entirely ineffective. Proactively bump it.

We know that's effective because it's enough to break an integration test that implicitly depended on no caching / cache misses, oops. Explicitly disable caching for that test.

Reviewed By: YousefSalama

Differential Revision: D60587372

fbshipit-source-id: d66210977f401a7afd6b87b0b75c848b43b708ab