Pass both movers to rewrite_commit
Summary:
## This stack
Implementing support to backsync changes that touch the submodule expansions.

# This diff
Backsyncing changes to the submodule expansion will require the small-to-large mover for both validation and the "compacting" of the submodule expansion.

Currently, we only pass the mover from the direction of the sync, so when backsyncing this is large to small.

This diff changes `rewrite_commit` to take `Movers` (an existing type that holds movers for both directions) instead of a single `Mover`, so that in the next diff we can pass the reverse mover when backsyncing changes to submodule expansion.

Most of the changes here are just replacing "mover" with "movers"...

Reviewed By: andreacampi

Differential Revision: D60176090

fbshipit-source-id: ae8bb41b28e326e58c14d86dab3c0878f4ffb1c5