walker: make walker tail logic generic over bulk fetcher type
Summary: To migrate the walker from using the changesets table, we'll need a generic way to use either `ChangesetBulkFetcher` or `CommitGraphBulkFetcher` to iterate over the changesets of a repo. This diff introduces `BulkFetcherOps` trait and implements it for `ChangesetBulkFetcher` while not changing the behaviour as a first step towards this.

Differential Revision: D60600041

fbshipit-source-id: b44d2e2395f0efea3abedab138a7fccf66f22e29