eagerepo: store pointers so we can read data by CasDigest
Summary: I want to make the eager store work as a CasClient, so it needs to be able to fetch file blobs and augmented tree blobs by CasDigest. Step 1 is to insert a pointer from CasDigest to the corresponding HgId of the file or tree. We store one extra byte indicating whether it is a file or a tree since we will need to strip the hg metadata off the start of the file content when reading.

Reviewed By: zzl0

Differential Revision: D60475565

fbshipit-source-id: c3bd5ca2533b9a2c5daedf52ed36eb65feb3fd0e