Extract current backsync logic to separate function
Summary:
## This stack
Implementing support to backsync changes that touch the submodule expansions.

# This diff
Currently we only support backsyncing changes to repos with submodule expansions if we confirm that the expansions are not being modified.

I'm extracting this logic into a separate function, because the proper support for backsyncing changes with submodule expansion will be introduced very soon in a diff above.

The new support will be wrapped under a JK that will fallback to the current behaviour, so let's wrap it under a function to keep the code cleaner.

Reviewed By: andreacampi

Differential Revision: D60663851

fbshipit-source-id: 0a38342bf328be9a69c4d69e3283be6e07dcc197