Replace large_repo_id with small_repo_id in SubmoduleExpansionData
Summary:
## This stack
Implementing support to backsync changes that touch the submodule expansions.

# This diff
This removes the `large_repo_id` field from SubmoduleExpansionData and adds `small_repo_id`.
large_repo_id was added before the `large_repo` field, which is an `InMemoryRepo` needed for validation and **also has the large repo id**.

The large repo id is used to determine the order of the sync.

I'm adding `small_repo_id` so we can log the correct target repo id when backsyncing (it's currently logging the large repo, which is wrong).

Reviewed By: andreacampi

Differential Revision: D60668289

fbshipit-source-id: b60eb95790f4ac6ca4738386c4d2326b5665bdbf