Kill submodule expansion support in repo_import
Summary:
## This stack
Implementing support to backsync changes that touch the submodule expansions.

# This diff
repo_import is an old tool to import git repos into Mononoke (not sure what's the overlap with gitimport0, but I'm not sure it's even used anymore...

It's one of the callsites of `rewrite_commit`, so while working on submodule expansion support throughout this entire year I had to keep updating it to make the compiler happy.

The reality is that submodule expansion is not needed on it. I've been passing partial information this whole time, which wouldn't work if we used...

It's best to be opinionated and just explicitly kill support for it.

If any need to support submodule expansions here arise, I can spend time understanding what repo_import is supposed to do and how it works, to provide the proper support.

Reviewed By: andreacampi

Differential Revision: D60668290

fbshipit-source-id: bc9bade082274d8bd70be4021bbecfe3b44fac36