commit_graph: implement add_many method
Summary: For batch inserting changesets into the commit graph, we currently rely on `CommitGraph::add_recursive` method which has extra logic to search for and fetch missing ancestors from the Changesets facet. As we are deleting the changesets facet we now need a batch insertion method that only inserts the changesets without any catchup logic.

Differential Revision: D60675069

fbshipit-source-id: ab2f1bee4b2477b32d15aa7d43ff26ad1207acbd