commit_graph_compat: use add_many instead of add_recursive when inserting new changesets
Summary: `add_recursive` has logic to search for missing ancestors in the Changesets facet when adding new changesets to the commit graph. This logic was only needed during the commit graph rollout which was more than a year ago now. The changesets and commit graph table have been in sync since then and this logic is no longer needed so let's switch to `add_many` instead.

Differential Revision: D60675259

fbshipit-source-id: 784cd5bd2d18f0e49845cd1dcb71e4f09412e0ad