Use read_master connection in GitPushRedirect facet
Summary: We use the `GitPushRedirect` facet in Mononoke Git server to determine if Mononoke is the source of truth for a repo. We want this to reflect the most immediate state and hence reading from replica is just not feasible. This diff changes the SQL connection from `read_connection` to `read_master_connection`.

Reviewed By: andreacampi

Differential Revision: D60671377

fbshipit-source-id: e2c41bac7d2731d334f5eecde0dd3e28ed60124f