Improve error logging for unexpected paths in expansion
Summary:
## This stack
Implementing support to backsync changes that touch the submodule expansions.

# This diff
I was debugging an unexpected behaviour in the SCS integration test in D60118876 (see comment) and noticed that this error handling could be better.

Instead of failing on the first unexpected file, let's get all of the unexpected files, log them and then return an error.
This will provide more useful debugging information in the logs.

Reviewed By: RajivTS

Differential Revision: D60760589

fbshipit-source-id: c90ba7ea2785e4d1a7dcf25eefa8bb862834facc