add additional error info on Windows
Summary:
# This diff

Adds additional error messaging to fbclone to give remediation advice to users on Windows.

# Context

Fbclone will consistently fail on Windows due to Edenfs or Sapling not being correctly installed. Users then post to the user group where they wait for us to respond. The remediation is usually pretty simple; there's a list of steps you can run through to try to get your host healthy again.

Therefore, we should add this info to the error message so that's it's easily accessible for users when the issue occurs.

We've found that this issue has occurred 106 times over the past month, so it's worthwhile to add this additional info to the error message. See https://fb.workplace.com/groups/edenfswindows/posts/1520441615538378/?comment_id=1537822017133671&reply_comment_id=1538297883752751

Reviewed By: kavehahmadi60

Differential Revision: D60523629

fbshipit-source-id: 729871dad68f74cff91b0df4849984d3c8fbbcfc