Add check for missing or corrupted SNAPSHOT
Summary:
# Context
Eden occasionally experiences corruptions due to missing files. Currently the detection and reporting of these errors is poor. This is part of an attempt to catch specific missing files and provide appropriate remidation steps

# This Diff
Adds a handler for the various RuntimeErrors that can occur when attempting to read the SNAPSHOT file. It also adds a new check if the snapshot file exists.

# Technical Details

NEW:
Add an existence check for the snapshot file before attempting to read it
Convert RuntimeErrors when trying to read hg to EdenCheckoutCorruption Problem. This provides additional detail and remediation instructions.

Reviewed By: genevievehelsel

Differential Revision: D60457714

fbshipit-source-id: a86ace15443149af9240286fdb198b47d464f0cd