Add checkout corruption problem
Summary:
# Context
Eden occasionally experiences corruptions due to missing files. Currently the detection and reporting of these errors is poor. This is part of an attempt to catch specific missing files and provide appropriate remidation steps

# This Diff
This diff adds a specific check for missing config.toml in the eden checkout. Before this would give a vague misconfigured checkout error and ask for help. Now it tells you to reclone. Next steps would be to try to automatically regenerate the config.toml

# Technical Details

A FileNotFound error is raised when attempting to read the config file if it does not exist. This is now specifically handled to raise an EdenCorruptionError, which is a catch-all for errors that require a reclone.

Reviewed By: MichaelCuevas

Differential Revision: D60392404

fbshipit-source-id: 98e3e865540dd45162d61cc75390acba505f68d5