Provide control over staleness over fetched SoT status
Summary: There are certain cases where we need the most up-to-date state for the SoT flag in Mononoke Git, e.g. when deciding whether to accept a push or not. However, we can't always ask for the most recent data cause that ends up directing the query to XDB master and every use-case relying on the master can cause the master to overload. This diff introduces an enum where the callsite can specify the level of recency that they need while fetching the SoT status for a given repo

Reviewed By: andreacampi

Differential Revision: D60768034

fbshipit-source-id: 25c2a8365f00af44138030fbc1e0a31a8e837e89