run hg debugnetworkdoctor in eden doctor
Summary:
# Context
Hg debugnetworkdoctor is a utility that diagnoses network issues. We already run this as part of eden rage. We now want to add this to eden doctor to add some additional signal on potential issues.

# This Diff
Add call in check_running_mount for network if scm type is HG and not fast
Add stub check function
Add initial problem NetworkDoctorProblem

# Technical Details
Not running on --fast due to debugnetwork taking an unknowable amount of time dependent on network conditions, though usually a couple seconds based on devserver testing. Still slow enough to want to avoid running with doctor, which is currently ~3 seconds total runtime

Reviewed By: genevievehelsel

Differential Revision: D60801935

fbshipit-source-id: 61d58b91673b94a6f7e3da7ea94bd2154735f43a