run hg debugnetworkdoctor returns 1 on error
Summary:
# Context
Hg debugnetworkdoctor is a utility that diagnoses network issues. We already run this as part of eden rage. We now want to add this to eden doctor to add some additional signal on potential issues.

# This Diff
Update return value to be 1 when there is an error diagnosed
# Technical Details
used next diff

Reviewed By: zzl0

Differential Revision: D60774516

fbshipit-source-id: 6e982fbec59ecdd4cbaf3b2a3b602a75deff397c