goto: fix eden goto with pending commits
Summary: Commands such as "debugimporttstack" can create commits and then check them out all in a single transction. For this to work on eden, the pending commit data needs to be flushed to disk so eden can see it when it performs the checkout operation. In Python we did this in EdenThriftClient.py, but we had no equivalent when Python calls into the Rust goto. Now we do.

Reviewed By: singhsrb

Differential Revision: D60805144

fbshipit-source-id: ee33055edd32225583d81cbba44ff231a968e812