cli: error out when command is not specified
Summary:
People have gotten accustomed to run `sl` in their repos. However, if for some reason there is an issue with the repo it can be both confusing and not very useful to print the help text.

This diff makes Sapling throw the same error as when other commands need a repo if `naked-default.no-repo` is not set. We can tweak the message on that error later if we want.

Reviewed By: muirdm

Differential Revision: D60626071

fbshipit-source-id: 9e9384ebaed6c5b0ec1305a22ebd82fe3efd8edd