refactor changeset writes to go through the CommitGraphWriter facet
Summary: This is a refactoring diff only. This diff makes all changeset inserts go through the `CommitGraphWriter` facet instead of the `Changesets` facet. The double writing logic is moved to the `CommitGraphWriter` facet so there should be no behaviour changes. This makes it easier to delete the `Changesets` facet in later diffs.

Reviewed By: gustavoavena

Differential Revision: D60834104

fbshipit-source-id: 06dec5fb9de65d2a1dcdc3922819fa3ead8bb3f5