Ensure that we allow atleast default number of additional changesets for hooks
Summary:
D57278832 introduced a configurable limit on the number of additional changesets on which hooks can be run. This change was to conditionally allow more than 200K commits. However, the current implementation of GitServer can end up restricting it to a much lower limit depending on the size of the push.

This diff changes that behaviour so that we atleast support hook execution on 200K commits and if the push is even bigger, we support that as well.

Differential Revision: D60902833

fbshipit-source-id: caa2f9a38447cfd4c0b7dc4088df22cd682b7405