copies: kill _filtercopies
Summary:
The docstring indicated it is only necessary for the copy tracing heuristics based algorithm, which I don't think we use anymore. Let's try removing it.

This method has been really annoying for cross-directory copy tracing since it tends to filter out copies that we actually want.

I examined the difference in all changed tests, and they all seem to now properly track a copy or rename where previously they did not. In certain "merge" cases we now properly remove a renamed path in the "other" as we follow copies to merge into the original path in "local".

Differential Revision: D60698882

fbshipit-source-id: 83381b17ccfdbccee3f7605d46371e92cda8bdcc