graft: fix another xdir copy trace case
Summary:
With this setup:

```
B  # B/B = B\n (copied from A)
|
A  # A/A = A\n

$ sl go $A
$ sl graft -r $B --from-path B --to-path A
```

the file exists on both sides (i.e. "A" on left, and "B" on right with "B" -> "A" mapping). The ancestor commit in this case is $A. The default file path for the ancestor file is "B" (obtained from "ungrafting" the "A" path). However "B" is not in the ancestor commit $A, so we look at copies and see it was copied from file "A". From here, previously we would ungraft "A" back to "B" (again) and use that as the ancestor file name, but that is not correct. Fix by not ungrafting and using "A" as the ancestor file name.

It seems like there should be a counter-example case where the ungraft was required, but I couldn't think it up.

Differential Revision: D60698879

fbshipit-source-id: a687aa0de235365d22b97a953e617844bcf030ae