hooks: add hook for invalid symlinks
Summary:
If an invalid symlink (e.g. longer than platform limit) is committed, it causes errors for clients who attempt to check it out. Sapling has some logic to ignore "suspect" symlinks in its "status" computation, but it is better to reject them in the server.

I kept the same logic that Sapling uses:
- max size of 1024
- disallow empty contents
- disallow newlines
- disallow null bytes

The most common reason for invalid symlinks getting committed is when a Windows user accidentally modifies a symlink, adding "normal" file contents. The client should probably have stricter validation for this case as well.

Differential Revision: D60859221

fbshipit-source-id: 4f5c43516019a181e091768bdb018b8cbdc7376b