cmdlib: make ChangesetArgs::resolve_changeset require exactly one changeset
Summary: `ChangesetArgs::resolve_changeset` currently returns an `Option` depending on whether a single changeset was provided or none. This requires all callsites to handle the `None` case, with most current callsites having misleading error messages that imply that the changeset was not found in the repo not that it was not provided. Let's make it not return an `Option` instead and require providing exactly one changeset instead.

Reviewed By: markbt

Differential Revision: D60900968

fbshipit-source-id: c8c0c5421ddbd7d6d49c3cf4ef74be44adde89c5