sharded_map_v2: process inlined nodes directly and avoid bounded traversal recursion
Summary:
Iteration over sharded maps uses bounded_traversal_ordered_stream.  Currently each step of the iteration, representing a node in the sharded map, is recursed upon through bounded traversal whether or not it is inlined.  For a manifest of 100k entries, this can result in 250k+ recursions.

Bounded traversal's accounting is expensive, so this means we spend a lot of time tracking these recursions.  However, they are not necessary: we only really need to use bounded traversal recursions in the case of things that might be async: i.e. nodes that are not stored inline.

In this diff, we change `into_prefix_entries_traverse` to add *all* entries for inlined nodes to the output.  This means we only use bounded traversal recursion for the async step of loading non-inlined nodes.  This greatly reduces the number of recursion steps from 250k+ to 200+.

Reviewed By: YousefSalama

Differential Revision: D60912702

fbshipit-source-id: f5b2eaec3fa6a63bd07c9b9abc7766307f027d48