sharded_map_v2: extract inner part of into_prefix_entries_impl
Summary:
Iteration over sharded maps uses bounded_traversal_ordered_stream.  Currently each step of the iteration, representing a node in the sharded map, is recursed upon through bounded traversal whether or not it is inlined.  For a manifest of 100k entries, this can result in 250k+ recursions.

Bounded traversal's accounting is expensive, so this means we spend a lot of time tracking these recursions.  However, they are not necessary: we only really need to use bounded traversal recursions in the case of things that might be async: i.e. nodes that are not stored inline.

We will change this so that we only use bounded traversal recursion for non-inline nodes.  In this diff, we first extract the inner part of the function to a separate function so that we can later on call it recursively for the inline case.

Reviewed By: YousefSalama

Differential Revision: D60912704

fbshipit-source-id: dff57dd7b0c165094ea02245065a881ac8218fa1