archive: don't write files in atomictemp mode
Summary:
The 'atomicfile' operation is not required for this usage.
Turning it off gives about a 40% speedup when the target of the archive
is an NFS volume.

https://fb.workplace.com/groups/mercurialusers/permalink/2822759524539576/

A future effort might attempt to do the metadata operations within
the single 'open'd file (though this is a trickier and less certain
win).

Reviewed By: muirdm

Differential Revision: D60797200

fbshipit-source-id: 3512198d9ba6fae894bf6cd99932f451c3c90f64