hg_augmented_manifests: lookup reused entries and partial maps in parallel when deriving HgAugmentedManifests
Summary: Right now when deriving `HgAugmentedManifests` we diff the corresponding hg manifest with its parents, then we go through the diff and make sequential sharded map lookup calls to fetch the reused entries and partial maps. This diff changes this to instead collect all path elements and afterwards make the lookup calls concurrently. Later on we can switch this to make a batch sharded map query for all lookups which should be even more optimized.

Reviewed By: andreacampi

Differential Revision: D60965495

fbshipit-source-id: 7322000b6e25b3f75394172272efc9a04bbe3ba0