Rename repo_a to small_repo in unit tests
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
In the unit tests we have a large repo and 3 git repos: A, B and C.

Repo A is used as the **small repo which is synced to the large repo**.

Let's make this more explicit and the overall tests less confusing by renaming everything from `repo_a` to `small_repo`.

Reviewed By: andreacampi

Differential Revision: D61014785

fbshipit-source-id: 63333076c357280c02b03050809491f1c10fc39d