Stop hardcoding expected changesets in unit tests
Summary:
## This stack
Continuing work to support backsyncing submodule extensions.

Most of it will be improving and adding tests to cover the edge cases.

## This diff
Stop hardcoding the expected changesets on the unit tests. This adds a lot of work when making any change to the test setup (e.g. next diff where I rename things, which modif all the bonsais).

we want to check if a mapping was inserted from the original changetset and the synced one, and we don't need to hardcode a value for that.

Reviewed By: mitrandir77

Differential Revision: D61014786

fbshipit-source-id: f6589b009eea3ec10a32fc89bf2df5dc7da3bacf