add journal as a depency of ScmStatusCache
Summary:
# Background
We have a plan to optimize `ScmStatusCache` by ignoring the file changes under `.hg` folder when counting the `Journal` sequence change.

# Tech Detail
In this diff, I'm adding `Journal` as a dependency of `ScmStatusCache` so we can access its flags and functions to implement `isSequenceValid()` function which will be used in follow-up diffs.

Reviewed By: kmancini

Differential Revision: D60856347

fbshipit-source-id: 5323d9c96bd7ae0aa97408d76d1621d639ca2386