record current working directory when computing diff
Summary:
# Background

We have a plan to optimize ScmStatusCache by ignoring the file changes under .hg folder when counting the Journal sequence change.

# Tech Detail

In this diff, I'm creating a new variable to hold a copy of the value of the current working directory. This will be used in next diff. We will detect if we are safe to continue using the `ScmStatusCache` by comparing this value with a cached working directory hash (to be added).

Reviewed By: kmancini

Differential Revision: D60857978

fbshipit-source-id: 863051869e95cead665089422513714e4fad9c31