sharded_map_v2: use quickcheck_async::tokio macro in quickcheck test
Summary: `test_sharded_map_v2_quickcheck` is a quickcheck test that uses a custom `quickcheck::TestHelper` implementation.  The way it's currently implemented makes it so that shrinking input doesn't work properly and so we don't get minimal test cases when a failure is found. Let's use the `quickcheck_async::tokio` macro instead which is simpler to use anyways.

Reviewed By: markbt

Differential Revision: D61018866

fbshipit-source-id: 4514a0e616eb58af03600ba006d5297ce6f61ec2