merge: hide some file merges from mergedrivers
Summary:
Originally D14195031 hid non-conflict merges from merge drivers by filtering out some copy information. However, the copy information is valid and should be used when merging. I deleted that logic in D60698882, but didn't pay close enough attention to the change in test-fb-ext-copytrace-mergedriver.t (thanks zzl0!).

In this commit I bring back the dont-expose-non-conflicts-to-merge-drivers with a different approach. If an "m" action doesn't actually look like a conflict, we don't add it to the merge state (this avoids the merge driver seeing it). We apply it as a "g" action (i.e. just get the contents and write to working copy). We keep the original action as an "m" so that recordupdates() later can record the copy information properly.

There could still be some non-conflicting merges that make it to merge drivers, but I think this should cover the common case.

Reviewed By: zzl0

Differential Revision: D61035082

fbshipit-source-id: 9441c82925a538d265a839f05d962a7e3bb95da9