Check if calls have sudo perms before running sudo commands
Summary:
# Context
Devservers have limited sudo permissions. Eden doctor uses sudo umount to unmount stale directories. This can cause problems on devservers

# This Diff
Adds a check to see if the user has umount permissions before running problem fixer. If the permissions are missing raise a RemediationError

# Technical Details
sudo -l XXX checks if the user has permission to run XXX. if so, it will return the path to the binary eg. /bin/umount
If not, it will return blank

Reviewed By: genevievehelsel

Differential Revision: D60973955

fbshipit-source-id: 7c78759bcba1f00eea4707ff90f569499c8027c6