commit cloud: get smartlog
Summary:
This is a new version of `get_smartlog`. The original commit in the cloud would call SCS for `get_stack_info` and iterate on the result. Instead, since we're in Mononoke, let's simply directly calculate the stacks and format the nodes.

In summary, this process fetches the head's bookmarks and remote bookmarks from the database. It then retrieves all the commits that form the stacks registered by the heads, compiles this information, and sends it serialized to the client.

Reviewed By: mzr

Differential Revision: D60239768

fbshipit-source-id: ed7ea3de95c16368b73697110bb710489dc527ba