Avoid race condition when updating repos
Summary:
For a sharded service or application, there are in total 3 ways a deep-sharded repo could get updated:
- SM makes an explicit request to add or remove a repo
- Task restarts and reads its local state and realizes it needs to boot up some list of repos
- Config updates requiring all the repos on disk to get reloaded

S439066 occurred as a result of 2 and 3 occurring together. When the task was busy restoring its known set of repos by loading them one by one, a config update happened. Before reloading the repos for the updated config, the config-updater tries to get list of repos that the server has, i.e. a snapshot. There is a time lag between taking the snapshot and reloading the repos and in the meanwhile a new repo could be added to the server. Since the config-reloader is not aware of this new repo, it silently discards it which results in the server losing the repo without any error or failure.

At this point, ShardManager and TW have different views of the task cause SM would still continue to believe that the repo is present on that task. This leads to requests getting redirected to this host which eventually fail causing the SEV.

This diff ensures that as part of config update we NEVER remove a repo but we can add or update. Note that there is a very small possibilty that as a result of this setup a repo could end up with the old config but that is definitely the lesser of the two evils here.

Reviewed By: andreacampi, mzr

Differential Revision: D60913304

fbshipit-source-id: 72452b485cad5f0f4991dc32c921420dd8aca4f9